Sync all filings in the last 14 days
There are some new forms, OCRA 305, that seem to have artificial "filed
at" dates, as they claim to be filed exactly at midnight. I wonder if
these were backfilled or something into the stream of filings? Three of
them didn't get downloaded by the download script in production, so I'm
assuming they were placed in the stream after already-existing forms.

Anyway, let's update the logic so it will save all the filings for the
last 14 days. Since it is time-consuming to download the contents, let's
only try to do that upon first download.